created a module validators.py for validating fight skills
